% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupDespliegue.R
\name{SetupDespliegue}
\alias{SetupDespliegue}
\title{Setup inicial de despliegue}
\usage{
SetupDespliegue(folder_datos, folder_scripts, mailpassword)
}
\arguments{
\item{folder_datos}{Path a folder donde esperamos guardar los datos.}

\item{folder_scripts}{Path a folder donde esperamos guardar los scripts.}
}
\value{
Variables guardadas localmente en sesion
}
\description{
Instala paquetes necesarios y crea variables para guardar y tomar scripts localmente.
}
\examples{
\dontrun{
# Usado por sus efectos secundarios
}
}
\author{
Eduardo Flores (Numbersmart)
}

